requirements.txt
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.0
flower==1.2.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.6.0
idna==3.4
iniconfig==2.0.0
kombu==5.2.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
passlib==1.7.4
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.6
pytest==7.2.2
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
redis==4.5.3
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
structlog==22.3.0
tomli==2.0.1
tornado==6.2
typing_extensions==4.5.0
uvicorn==0.21.1
vine==5.0.0
wcwidth==0.2.6
